+++ backend-server/application/src/main/java/com/apitable/workspace/vo/FieldRole.java
@@ -54,7 +54,7 @@ public class FieldRole {
     @Schema(description = "The number of members of the department. It is returned when the type "
         + "is department", example = "3")
     @JsonSerialize(nullsUsing = NullNumberSerializer.class)
-    private Integer memberCount;
+    private Long memberCount;
 
     @Schema(description = "Head portrait, returned when the type is member",
         example = "https://www.apitable.com/image.png")
